+++ airwindows/src/Acceleration.cpp
@@ -64,7 +64,7 @@ void _airwindowsAlgorithm::_kernel::render( const Float32* inSourceP, Float32* i
 	overallscale /= 44100.0f;
 	overallscale *= GetSampleRate();
 
-	Float32 intensity = pow(GetParameter( kParam_One ),3)*(32/overallscale);
+	Float32 intensity = powf(GetParameter( kParam_One ),3)*(32/overallscale);
 	Float32 wet = GetParameter( kParam_Two );
 	//removed unnecessary dry variable
